+++ Rock-master/src/main/java/com/movie/rock/movie/data/repository/MovieGenresRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface MovieGenresRepository extends JpaRepository<MovieGenresEntity, Integer> {
+public interface MovieGenresRepository extends JpaRepository<MovieGenresEntity, Long> {
     List<MovieGenresEntity> findByMovieMovieId(Long movieId);
 }
